feat(account): wire cash balance settlement into order placement
- Implement event bus (EDA) for decoupling
  * ledger/balance updates from order and auth flows
- Rename UserBalance -> CashAccount across models/services, enforce true 1:1 (prev: 1 : many), unique constraint on user_id, add DB-level CHECK constraints on balance invariants, new migration
- Fix two pre-existing bugs in an earlier migration that were blocking the chain: a bad DDL drop order, and a missing type-creation step in an enum rename (both directions)
- Fix ledger entry to update cash balance in cash account on order settle
- Add AccountService + account related endpoints
- Add buying power check (BalanceStatus); add matching withdrawable check in account service
- Switch trade quantity/price fields from float to Decimal
- Drop redundant db.commit()/db.refresh() calls where exists (get_db() auto-commits; SQLAlchemy 2.0 eager_defaults populates INSERT-time server defaults on flush)

+++ schemas/trade.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from typing import Literal, Optional
 
 from pydantic import BaseModel
@@ -8,9 +9,9 @@ class TradeOrderRequest(BaseModel):
     instrument_type: Literal["stock", "etf", "fund"]
     side: Literal["buy", "sell"]
     order_type: Literal["market", "limit", "stop", "stop_limit"]
-    quantity: float
-    limit_price: Optional[float] = None
-    stop_price: Optional[float] = None
+    quantity: Decimal
+    limit_price: Optional[Decimal] = None
+    stop_price: Optional[Decimal] = None
     time_in_force: Literal["day", "gtc", "ioc"] = "day"
 
 
@@ -20,5 +21,5 @@ class OrderConfirmation(BaseModel):
     ticker: str
     side: str
     order_type: str
-    quantity: float
-    estimated_total: float
+    quantity: Decimal
+    estimated_total: Decimal
